requirements.txt
streamlit~=1.37.1
streamlit-local-storage~=0.0.23
groq~=0.9.0
python-dotenv~=1.0.1
requests~=2.32.3
